# turma-do-gatil-back/api-definition.yaml
openapi: 3.0.3
info:
  title: Turma do Gatil - API
  description: API REST para gerenciamento do sistema Turma do Gatil - Sistema de adoção de gatos
  version: 1.0.0
  contact:
    name: Equipe Turma do Gatil
    email: contato@turmadogatil.com
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
  termsOfService: https://turmadogatil.com/terms
servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento
...
paths:
  /api/cats:
    get:
      tags:
        - Cats
      summary: Lista todos os gatos
      description: Retorna uma lista paginada de gatos com filtros opcionais
      parameters:
        - name: page
          in: query
          description: Número da página (inicia em 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Quantidade de itens por página
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Campo para ordenação
          required: false
          schema:
            type: string
            default: name
        - name: sortDir
          in: query
          description: Direção da ordenação
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: name
          in: query
          description: Filtro por nome do gato
          required: false
          schema:
            type: string
        - name: color
          in: query
          description: Filtro por cor do gato
          required: false
          schema:
            $ref: '#/components/schemas/Color'
        - name: sex
          in: query
          description: Filtro por sexo do gato
          required: false
          schema:
            $ref: '#/components/schemas/Sex'
        - name: adopted
          in: query
          description: Filtro por status de adoção
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de gatos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCat'
    post:
      tags:
        - Cats
      summary: Cria um novo gato
      description: Adiciona um novo gato ao sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatRequest'
      responses:
        '201':
          description: Gato criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
        '400':
          description: Dados inválidos
  /api/cats/{id}:
    get:
      tags:
        - Cats
      summary: Busca gato por ID
      description: Retorna um gato específico pelo seu ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do gato
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gato encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
        '404':
          description: Gato não encontrado
    put:
      tags:
        - Cats
      summary: Atualiza um gato
      description: Atualiza os dados de um gato existente
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do gato
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatRequest'
      responses:
        '200':
          description: Gato atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
        '404':
          description: Gato não encontrado
        '400':
          description: Dados inválidos
    delete:
      tags:
        - Cats
      summary: Remove um gato
      description: Remove um gato do sistema
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do gato
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Gato removido com sucesso
        '404':
          description: Gato não encontrado
  /api/cats/needing-sterilization:
    get:
      tags:
        - Cats
      summary: Busca gatos que precisam de castração
      description: |
        Retorna uma lista de gatos que estão aptos para castração baseado nas seguintes regras:
        
        **Critérios de Elegibilidade:**
        - Gato deve ter pelo menos **90 dias de vida**
        - Gato **não deve ter sido castrado** (status COMPLETED)
        - Gato **não deve ter castração agendada** (status SCHEDULED)
        - Apenas gatos **não adotados** são considerados
        
        **Status Retornados:**
        - `ELIGIBLE`: Gato com 90-179 dias, apto para castração
        - `OVERDUE`: Gato com 180+ dias, castração atrasada
        
        **Informações Retornadas:**
        - Dados básicos do gato (nome, cor, sexo, foto)
        - Idade exata em dias
        - Status de prioridade para castração
      responses:
        '200':
          description: Lista de gatos que precisam de castração retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatSterilizationStatus'
        '500':
          description: Erro interno do servidor ao processar a solicitação
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /api/cats/sterilization-stats:
    get:
      tags:
        - Cats
      summary: Obter estatísticas de castração
      description: |
        Retorna as estatísticas resumidas sobre gatos que precisam de castração:
        
        **Informações Retornadas:**
        - `eligibleCount`: Quantidade de gatos elegíveis para castração (90-179 dias)
        - `overdueCount`: Quantidade de gatos com castração atrasada (180+ dias)
        - `totalNeedingSterilization`: Total de gatos que precisam de castração
        
        **Critérios Aplicados:**
        - Apenas gatos não adotados
        - Gatos com pelo menos 90 dias de vida
        - Excluídos gatos já castrados ou com castração agendada
      responses:
        '200':
          description: Estatísticas de castração retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SterilizationStats'
        '500':
          description: Erro interno do servidor ao processar a solicitação
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /api/cats/dashboard-summary:
    get:
      tags:
        - Cats
      summary: Obter resumo do dashboard
      description: |
        Retorna um resumo completo para dashboards com as principais informações do sistema:
        
        **Informações Incluídas:**
        - `availableCats`: Lista dos primeiros 10 gatos disponíveis para adoção (não adotados)
        - `pendingSterilizations`: Lista das primeiras 10 castrações pendentes (status SCHEDULED)
        - `registeredAdopters`: Lista dos primeiros 10 adotantes cadastrados mais recentes
        
        **Características:**
        - Ideal para dashboards e telas de resumo
        - Fornece uma visão geral rápida do sistema
        - Dados limitados para performance otimizada
        - Gatos ordenados por nome (A-Z)
        - Castrações ordenadas por data agendada (mais próximas primeiro)
        - Adotantes ordenados por data de registro (mais recentes primeiro)
        
        **Utilidade:**
        - Monitoramento rápido do status do abrigo
        - Identificação de prioridades de atendimento
        - Visão consolidada para tomada de decisões
      responses:
        '200':
          description: Resumo do dashboard retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
              example:
                availableCats:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Mimi"
                    color: "ORANGE"
                    sex: "FEMALE"
                    birthDate: "2024-01-15T00:00:00"
                    shelterEntryDate: "2024-02-01T00:00:00"
                    photoUrl: "https://example.com/photos/mimi.jpg"
                    adopted: false
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    name: "Felix"
                    color: "BLACK"
                    sex: "MALE"
                    birthDate: "2024-02-10T00:00:00"
                    shelterEntryDate: "2024-02-15T00:00:00"
                    photoUrl: "https://example.com/photos/felix.jpg"
                    adopted: false
                pendingSterilizations:
                  - id: "660e8400-e29b-41d4-a716-446655440000"
                    catId: "550e8400-e29b-41d4-a716-446655440000"
                    cat: "Mimi"
                    photoUrl: "https://example.com/photos/mimi.jpg"
                    sterilizationDate: "2024-03-15T10:30:00"
                    status: "SCHEDULED"
                    notes: "Primeira castração do mês"
                registeredAdopters:
                  - id: "770e8400-e29b-41d4-a716-446655440000"
                    firstName: "João"
                    lastName: "Silva"
                    birthDate: "1990-05-15T00:00:00"
                    cpf: "123.456.789-00"
                    phone: "(11) 99999-9999"
                    email: "joao@email.com"
                    address: "Rua das Flores, 123"
                    registrationDate: "2024-02-15T14:30:00"
        '500':
          description: Erro interno do servidor ao processar a solicitação
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...
  /api/sterilizations/cat/{catId}:
    get:
      tags:
        - Sterilizations
      summary: Busca esterilizações por gato
      description: Retorna todas as esterilizações de um gato específico
      parameters:
        - name: catId
          in: path
          required: true
          description: ID único do gato
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de esterilizações do gato
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sterilization'
        '404':
          description: Gato não encontrado
...
components:
  schemas:
    Cat:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único do gato
        name:
          type: string
          description: Nome do gato
        color:
          $ref: '#/components/schemas/Color'
        sex:
          $ref: '#/components/schemas/Sex'
        birthDate:
          type: string
          format: date-time
          description: Data de nascimento do gato
        shelterEntryDate:
          type: string
          format: date-time
          description: Data de entrada no abrigo
        photoUrl:
          type: string
          description: URL da foto do gato
        adopted:
          type: boolean
          description: Status de adoção do gato (compatibilidade)
        adoptionStatus:
          $ref: '#/components/schemas/CatAdoptionStatus'
    CatRequest:
      type: object
      required:
        - name
        - color
        - sex
        - birthDate
        - shelterEntryDate
      properties:
        name:
          type: string
          description: Nome do gato
        color:
          $ref: '#/components/schemas/Color'
        sex:
          $ref: '#/components/schemas/Sex'
        birthDate:
          type: string
          format: date-time
          description: Data de nascimento do gato
        shelterEntryDate:
          type: string
          format: date-time
          description: Data de entrada no abrigo
        photoUrl:
          type: string
          description: URL da foto do gato
        adopted:
          type: boolean
          default: false
          description: Status de adoção do gato (compatibilidade)
        adoptionStatus:
          $ref: '#/components/schemas/CatAdoptionStatus'
          default: NAO_ADOTADO
    Adopter:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único do adotante
        firstName:
          type: string
          description: Primeiro nome do adotante
        lastName:
          type: string
          description: Sobrenome do adotante
        birthDate:
          type: string
          format: date-time
          description: Data de nascimento do adotante
        cpf:
          type: string
          description: CPF do adotante
        phone:
          type: string
          description: Telefone do adotante
        email:
          type: string
          format: email
          description: Email do adotante
        address:
          type: string
          description: Endereço do adotante
        registrationDate:
          type: string
          format: date-time
          description: Data de cadastro do adotante
...
    Sterilization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único da esterilização
        catId:
          type: string
          format: uuid
          description: ID do gato
        cat:
          type: string
          description: Nome do gato
        photoUrl:
          type: string
          format: uri
          description: URL da foto do gato
        sterilizationDate:
          type: string
          format: date-time
          description: Data da esterilização
        status:
          $ref: '#/components/schemas/SterilizationStatus'
        notes:
          type: string
          description: Observações sobre a esterilização
    Color:
      type: string
      enum:
        - WHITE
        - BLACK
        - GRAY
        - BROWN
        - ORANGE
        - MIXED
        - CALICO
        - TABBY
        - SIAMESE
        - OTHER
      description: Cor do gato
    Sex:
      type: string
      enum:
        - MALE
        - FEMALE
      description: Sexo do gato
...
    CatAdoptionStatus:
      type: string
      enum:
        - NAO_ADOTADO
        - EM_PROCESSO
        - ADOTADO
      description: Status de adoção do gato
    SterilizationStatus:
      type: string
      enum:
        - SCHEDULED
        - COMPLETED
        - CANCELED
      description: Status da esterilização
    CatSterilizationStatus:
      type: object
      description: DTO que representa um gato com informações sobre seu status de castração
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do gato
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Nome do gato
          example: "Mimi"
        color:
          $ref: '#/components/schemas/Color'
        sex:
          $ref: '#/components/schemas/Sex'
        birthDate:
          type: string
          format: date-time
          description: Data de nascimento do gato
          example: "2024-03-15T10:30:00"
        shelterEntryDate:
          type: string
          format: date-time
          description: Data de entrada no abrigo
          example: "2024-04-01T09:00:00"
        photoUrl:
          type: string
          description: URL da foto do gato
          example: "https://example.com/photos/mimi.jpg"
        ageInDays:
          type: integer
          description: Idade do gato em dias
          example: 120
        sterilizationStatus:
          $ref: '#/components/schemas/SterilizationEligibilityStatus'
    SterilizationEligibilityStatus:
      type: string
      enum:
        - ELIGIBLE
        - OVERDUE
      description: Status de elegibilidade para castração
    SterilizationStats:
      type: object
      description: DTO que representa as estatísticas de castração dos gatos
      properties:
        eligibleCount:
          type: integer
          format: int64
          description: Quantidade de gatos elegíveis para castração (90-179 dias)
          example: 15
        overdueCount:
          type: integer
          format: int64
          description: Quantidade de gatos com castração atrasada (180+ dias)
          example: 8
        totalNeedingSterilization:
          type: integer
          format: int64
          description: Total de gatos que precisam de castração
          example: 23
    DashboardSummary:
      type: object
      description: DTO que contém um resumo do dashboard com gatos disponíveis, castrações pendentes e adotantes
      properties:
        availableCats:
          type: array
          description: Lista de gatos disponíveis para adoção (máximo 10)
          items:
            $ref: '#/components/schemas/Cat'
          example:
            - id: "550e8400-e29b-41d4-a716-446655440000"
              name: "Mimi"
              color: "ORANGE"
              sex: "FEMALE"
              birthDate: "2024-01-15T00:00:00"
              shelterEntryDate: "2024-02-01T00:00:00"
              photoUrl: "https://example.com/photos/mimi.jpg"
              adopted: false
        pendingSterilizations:
          type: array
          description: Lista de castrações pendentes/agendadas (máximo 10)
          items:
            $ref: '#/components/schemas/Sterilization'
          example:
            - id: "660e8400-e29b-41d4-a716-446655440000"
              catId: "550e8400-e29b-41d4-a716-446655440000"
              cat: "Mimi"
              photoUrl: "https://example.com/photos/mimi.jpg"
              sterilizationDate: "2024-03-15T10:30:00"
              status: "SCHEDULED"
              notes: "Primeira castração do mês"
        registeredAdopters:
          type: array
          description: Lista de adotantes cadastrados mais recentes (máximo 10)
          items:
            $ref: '#/components/schemas/Adopter'
          example:
            - id: "770e8400-e29b-41d4-a716-446655440000"
              firstName: "João"
              lastName: "Silva"
              birthDate: "1990-05-15T00:00:00"
              cpf: "123.456.789-00"
              phone: "(11) 99999-9999"
              email: "joao@email.com"
              address: "Rua das Flores, 123"
              registrationDate: "2024-02-15T14:30:00"
...
    PageCat:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Cat'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
        first:
          type: boolean
        numberOfElements:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
    Pageable:
      type: object
      properties:
        sort:
          type: object
          properties:
            sorted:
              type: boolean
            unsorted:
              type: boolean
        pageNumber:
          type: integer
        pageSize:
          type: integer
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean